test(gateway-client): extract metrics test
