+++ crates/gateway-client/Cargo.toml
@@ -7,13 +7,14 @@ rust-version = "1.62"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-test-utils = ["dep:mockall"]
+test-utils = ["dep:http", "dep:mockall", "dep:serde_json", "dep:tokio", "dep:warp"]
 
 [dependencies]
 anyhow = { workspace = true }
 async-trait = "0.1.59"
 bytes = "1.3.0"
 futures = { version = "0.3", default-features = false, features = ["std"] }
+http = { version = "0.2.8", optional = true }
 metrics = "0.20.1"
 mockall = { version = "0.11.3", optional = true }
 pathfinder-common = { path = "../common" }
@@ -21,24 +22,23 @@ pathfinder-retry = { path = "../retry" }
 pathfinder-serde = { path = "../serde" }
 reqwest = { version = "0.11.13", features = ["json"] }
 serde = { version = "1.0.149", features = ["derive"] }
+serde_json = { version = "1.0.89", features = ["arbitrary_precision", "raw_value"], optional = true }
 starknet-gateway-types = { path = "../gateway-types" }
+tokio = { workspace = true, features = ["macros", "test-util"], optional = true }
 tracing = "0.1.37"
+warp = { version = "0.3.3", optional = true }
 
 [dev-dependencies]
 assert_matches = "1.5.0"
 base64 = "0.13.1"
 flate2 = "1.0.25"
-http = "0.2.8"
 lazy_static = "1.4.0"
 pathfinder-common = { path = "../common", features = ["test-utils"] }
 pathfinder-serde = { path = "../serde" }
 pretty_assertions = "1.3.0"
 reqwest = { version = "0.11.13", features = ["json"] }
-serde_json = { version = "1.0.89", features = ["arbitrary_precision", "raw_value"] }
 stark_hash = { path = "../stark_hash" }
 starknet-gateway-test-fixtures = { path = "../gateway-test-fixtures" }
 test-log = { version = "0.2.11", default-features = false, features = ["trace"] }
-tokio = { workspace = true, features = ["macros", "test-util"] }
 tracing-subscriber = { version = "0.3.16", features = ["env-filter"] }
-warp = "0.3.3"
 zstd = "0.12"
