+++ crates/pathfinder/src/monitoring.rs
@@ -96,6 +96,8 @@ mod tests {
         // Other concurrent tests could be setting their own recorders
         let guard = RecorderGuard::lock(recorder);
 
+        // We don't care about the recorder being a singleton as the counter name here does not
+        // interfere with any other "real" counter registered in pathfinder or other tests
         let counter = metrics::register_counter!("x");
         counter.increment(123);
